split different client implementation in different file + add empty http client
